Add a die() function
For now, it will only be defined if no utility, function or alias
already exists by that name. Like the exit builtin, it is able to
preserve the value of $? but will never exit with a status of 0. For
example, the following will preserve the exit status of failing_command.

failing_command || die "gadzooks"

Whereas the following will ignore the value of $? - being that it is 0 -
and instead exit with a status of 1.

if prevailing_command; then
	die "fiddlesticks"
fi

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Bug: https://bugs.gentoo.org/878505