Relax parameter validation for various functions
Following some deliberation over the matter, I have concluded that
calling die() for the handling of invalid parameters ought not to be
considered as a matter of course. As such, this commit retains the
existing diagnostics for the functions listed below, while modifying
them so as to no longer attempt to exit the shell.

- eend
- eqatag
- esyslog
- ewend
- is_older than
- veend
- vewend
- yesno

Signed-off-by: Kerin Millar <kfm@plushkava.net>