Add the srandom() function
This is based on the behaviour of the special SRANDOM variable in bash.
It should be noted that sh is capable of bitwise arithmetic. For
instance, it is possible to clamp the number in such a way that it does
not exceed the minimum possible value of RAND_MAX (32767).

n=$(srandom) && : $(( n >>= 17 ))

Signed-off-by: Kerin Millar <kfm@plushkava.net>