Deprecate RC_NOCOLOR and the equivalent use of the positional parameters
Firstly, the "." builtin is not specified to support the passing of
arguments; such is a bashism. It is rather presumptuous to act as if the
present contents of the positional parameters are for the benefit of
gentoo-functions at the time of its initialisation.

Secondly, there exists a de-facto standard for suppressing colored
output, which is to define NO_COLOR as a non-empty string, per
https://no-color.org.

Signed-off-by: Kerin Millar <kfm@plushkava.net>