Add an experimental module for staging new ideas
The idea is to be able to write new functions without having to
initially commit to their being in the core library or, indeed, any of
the other modules. Experimental functions may be promoted or simply
dropped without warning if it is decided that they are insufficiently
useful to merit incorporation.

This initial commit defines the is_interactive() and prepend_ts()
functions.

Signed-off-by: Kerin Millar <kfm@plushkava.net>