Make _select_by_mtime() work correctly for paths read from STDIN
The _select_by_mtime() function is called by both newest() and oldest().
Pathnames may be specified as positional parameters or as NUL-separated
records to be read from the standard input. Unfortunately, the latter
interface does not work at all. Rectify this by checking whether the
number of parameters is greater then 0, rather than greater than or
equal to 0.

Also, extend the existing test case in such a way that the interface in
question is tested.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>