test-functions: refactor the newest() test case
Rework the test case for the newest() function in accordance with the
recently added test case for the oldest() function. The resulting code
is more pleasant to read and maintain. In doing so, an obscure bug has
been addressed. Hitherto, an empty NUL-terminated record had erroneously
being conveyed to newest() for just one of the 28 individual sub-tests
being conducted.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>