test-functions: avoid unspecified behaviour in test_quote_args()
In test_quote_args(), there is the following code.

fmt=$(printf '\%o' "$i")

However, the behaviour of the <backslash> character followed by the
<number-sign> character is unspecified. Since it is intended to be taken
as a literal backslash, fix it by writing it as thus.

fmt=$(printf '\\%o' "$i")

Doing so addresses a spurious test failure where using the loksh shell.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>