test-functions: test for simple commands persisting environmental changes
Some implementations allow for alterations made to the execution
environment to persist beyond the scope of a simple command. Consider
loksh as a case in point.

$ f() { :; }
$ unset LEAKED
$ LEAKED=1 /bin/true; echo "LEAKED = $LEAKED"
LEAKED =
$ LEAKED=1      cmd2; echo "LEAKED = $LEAKED"
LEAKED = 1

Strictly speaking, such behaviour is permitted. The Shell Command
Language specification states:

"""
If the command name is a function that is not a standard utility
implemented as a function, variable assignments shall affect the current
execution environment during the execution of the function. It is
unspecified:

- Whether or not the variable assignments persist after the completion
  of the function

- Whether or not the variables gain the export attribute during the
  execution of the function

- Whether or not export attributes gained as a result of the variable
  assignments persist after the completion of the function (if variable
  assignments persist after the completion of the function)
"""

Unfortunately, loksh elects not to be aligned with the practices of the
overwhelming majority of implementations in this regard. For now, have
test-functions detect and abort for shells that go against the grain. I
shall consider reviewing and adapting gentoo-functions to account for
such unspecified behaviour but it is not an immediate priority.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>