Use an entropy pool in srandom(), even if the shell has forked
Presently, there are two srandom() implementations that do not require
bash, one of which is intended for use with mksh and the other of which
is intended for the various other implementations of sh(1). Both of
these implementations are capable of maintaining an entropy pool, which
markedly enhances performance for repeated invocations of the function.
However, the pool cannot be effectively utilised in cases where the
shell has forked.

$ srandom              # initialises the pool
$ srandom              # reads from the now-initialised pool
$ ( srandom )          # may fork, rendering the pool rather ineffective
$ ( srandom; srandom ) # ditto, despite the consecutive calls

This commit addresses the discrepancy by keeping track of whether the
pool has been populated on a per-PID basis. Consider the following
benchmark, in which the loop is forced to execute within a subshell
environment.

(
  i=0
  while [ $((i+=1)) -le 30000 ]; do srandom; done >/dev/null
  /bin/true
)

As conducted with mksh 59c on a system with a 2nd generation Intel Xeon,
I obtained the following figures.

BEFORE

real    3m8.857s
user    2m57.276s
sys     0m59.511s

AFTER

real    1m24.047s
user    1m6.435s
sys     0m19.565s

As conducted with dash on the same system, I obtained the following
figures.

BEFORE

real    0m52.056s
user    1m2.913s
sys     0m18.143s

AFTER

real    0m12.887s
user    0m12.521s
sys     0m1.016s

Signed-off-by: Kerin Millar <kfm@plushkava.net>