Reduce the two non-bash srandom() implementations to just one
The implementation of srandom() that was written with mksh first and
foremost in mind is no longer as slow as it was. I decided to benchmark
30,000 iterations of both of the non-bash implementations with varying
maximal pool sizes. The results are beneath. Note that both "dash/1" and
"mksh/1" refer to the mksh-targeting implementation.

Pool Size  dash/1   dash/2    mksh/1
     48 B   6.67s    5.57s    58.84s
     64 B   5.39s    4.78s    58.20s
     96 B   5.49s    4.36s    58.13s
    128 B   5.87s    4.63s    59.94s
    160 B   5.93s    5.46s    64.64s

These figures demonstrate that the optimal pool size is roughly between
64 and 96 bytes, and that the performance of both implementations is now
comparable. In addition to testing Linux (6.6) on x86_64 hardware, I
experimented with the pool size on macOS Sonoma (using an Apple M1 CPU)
and found a value of 64 to be close to optimal.

In view of these findings, have _collect_entropy() collect 64 bytes at a
time and remove the marginally faster implementation. That is, the one
that depended on being able to perform arithmetic on a number as high as
2^32-1 without overflowing.

Additionally, increase the maximum number of times that the remaining
implementation tries to find a suitable sequence of hex digits from 2 to
3. Finally, remove the overflow check, for it is no longer required.

Signed-off-by: Kerin Millar <kfm@plushkava.net>