Disable shellcheck SC2153
SC2153 is informational in nature and triggers only for environment
variables (all uppercase variables) whose names are similar to others
and for which no explicit assignment can be observed. In the case of
gentoo-functions, it was being raised as a result of KSH_VERSION and
YASH_VERSION being expanded. In other words, it is a nuisance.

Signed-off-by: Kerin Millar <kfm@plushkava.net>