Have whenceforth() work around a word splitting bug in OpenBSD sh
Consider the case where IFS consists of a single character whose value
is neither <space>, <tab> nor <newline>. The following example employs
the colon, since it is the character that the whenceforth() function
relies upon during word splitting.

$ bash -c 'IFS=":"; path=":"; set -- $path; echo "$# ${1@Q}"'
1 ''

The result is very much as expected because the colon in path serves as
a terminator for an empty field. Now, let's consider how many fields are
produced in OpenBSD sh as a consequence of word splitting.

$ sh -c 'IFS=":"; path=":"; set -- $path; echo "$#"'
0

For the time being, work around it by having whenceforth() repeat the
field terminator for the affected edge cases, which are two in number.
With this change, the test suite is now able to pass for:

- loksh 7.5
- oksh 7.5
- sh (OpenBSD 7.5)

Signed-off-by: Kerin Millar <kfm@plushkava.net>