Select an srandom() implementation upon being sourced
Presently, srandom() is defined as a self-redeclaring function, with the
the actual implementation being chosen upon its first invocation.
However, given that the function is highly likely to be invoked from a
command substitution, it isn't sensible to defer this choice.

Consider the following scenario.

  num=$(srandom) # capture a random number
  num=$(srandom) # capture a random number again

For both calls, srandom() will choose an implementation before
re-declaring and calling itself. That is, the initially chosen
implementation won't endure, since it only affects the subshell
environment that is incurred by the command substitution.

Address this minor issue by instead choosing an implementation at the
time that gentoo-functions is sourced.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>