Simplify the portable srandom() implementation
Presently, there are two implementations of the srandom() function, one
of which is intended for bash while the other is intended for sh. The
latter implementation employs a pattern to search for 8 consecutive hex
digits in the entropy pool, where the first is between 0 and 7.

Simplify matters by unconditionally consuming 8 hex digits from the pool
before employing a case statement that clamps the first digit, provided
that it is greater than 0x7. The resulting code is more pleasant to read
and maintain. What's more, it improves the performance in dash by
approximately 10%.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>