Refrain from negating the break builtin
Presently, there are two instances in which "! break" is used to break a
loop, with the expectation that the resulting status be non-zero. That
is, given that the exit status of break is 0, the act of negating it is
supposed to produce a status of 1. Unfortunately, the various sh(1)
implementations cannot be relied upon to behave in this manner.

$ bash -c 'for x in 1; do ! break; done; echo "$?"'
1
$ dash -c 'for x in 1; do ! break; done; echo "$?"'
0

Address this issue by using return instead.

Link: https://www.mail-archive.com/austin-group-l@opengroup.org/msg12302.html
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>