Work around python's pty module being unable to report its dimensions
Presently, portage sets up a pseudo-terminal by way of its
_create_pty_or_pipe() utility function. It makes use of Python's pty
module, which is unable to properly report its dimensions. Thus far,
those among the Python community that have recognised this issue appear
to have settled for the quintessentially pythonic approach of sweeping
it under the rug.

For the time being, have the _update_columns() function check whether
the shell appears to be a subprocess of portage and, if so, defer to the
existing value of the COLUMNS variable under the presumption that
portage has exported it. This is the very same 'solution' that is
employed by some of the bash functions that belong to portage.

See-also: 6f6440c538efb2edbcdfe298521252b6510a1a80
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>