Initialise the hr variable in the hr() function
The hr() function composes a horizontal rule by appending characters to
the 'hr' variable, which is never initialised. One might think that it
matters not, given that the variable is declared as function-local.
Unfortunately, it does matter.

$ testcase='var=leaked; f() { local var; echo "var=${var-unset}"; }; f'
$ set -- bash 'busybox sh' ksh mksh yash dash
$ for sh; do printf '%s: ' "$sh"; $sh -c "$testcase"; done
bash: var=unset
busybox sh: var=unset
ksh: var=unset
mksh: var=unset
yash: var=unset
dash: var=leaked

It can be seen here that dash is an outlier. Such are the perils of
depending upon a feature whose behaviour is undefined by the Shell
Command Language specification. Address this issue by ensuring that the
'hr' variable is always initialised as the null string.

Fixes: 8a83cf36a847fbd32990d3590bfd22a1516af898
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>