Don't let _update_columns() restore IFS as '' where previously unset
One of the methods by which the _update_columns() function is able to
determine the dimensions of the terminal is to execute the stty(1)
utility. Should it elect to do so, it backs up the value of the 'IFS'
variable to one named 'genfun_ifs' before setting 'IFS' appropriately.
Then, after stty(1) has been executed, the prior value of 'IFS' is
restored from 'genfun_ifs'. All of which is fundamentally incorrect.

$ unset -v IFS
$ echo "IFS=${IFS-unset}"
IFS=unset
$ f() { old_ifs=$IFS; IFS=$old_ifs; }; f
$ echo "IFS=${IFS-unset}"
IFS=

Note how, despite IFS having been unset, the function 'restores' it by
assigning the null string. But these two states are not equivalent. The
specification exacts the following two requirements.

"If the IFS variable is set and has an empty string as its value, no
field splitting shall occur."

"If the IFS variable is unset, then for the purposes of this section,
but without altering the value of the variable, its value shall be
considered to contain the three single-byte characters <space>, <tab>,
and <newline> from the portable character set, all of which are IFS
white-space characters."

Address this issue by jettisoning the 'genfun_ifs' variable and
declaring 'IFS' as a local variable instead.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>